shim: Extend section names to their long representation
Some section names are shortened up in some cases. The names of those
sections is read instead by a string formatted such as "/<integer>". The
<integer> is the offset of the strings table in which the long name of
the section is stored. The strings table is right after the symbols
table, so we can figure out where it starts knowing the beginning of the
symbols table, the numer of symbols and that each symbol is 18 bytes.

Finally, we read the long section name from the given offset of the
file, until we encounter a null character.

This extends the Shim struct to also contain the path to the shim binary
file. That way, we can read at specific offsets to extract long section
names.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>