shim: Select previous/latest sbatlevel policy
Two policies are shipped by default in the .sbatlevel section: previous
and latest. The offset to each of them can be found in the last two u32
of the .sbatlevel section data header. Then, each policy is delimited by
a null character.

This commit adds a mechanism to filter out whatever else and only get
the piece of the .sbatlevel that matters. By default, the "previous"
policy is selected, so that's why that's been hardcoded for now in
src/main.rs.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>