secureboot: Model secureboot state as a boolean
This makes the code rely on a boolean flag instead of the efivars
SecureBoot variable to know whether secureboot was enabled or not.

This makes it easier to write some of the following conditions that
depend on that value, as well as integrating other file format inputs
that store secureboot state, such as uefi or azdisk profiles, which do
not carry any information about the state of secure boot enablement.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>