shim: Refactor pe file specifics
The shim::Shim structure implemented a bunch of methods that were not
specific to shim but to pe files. In other words, the Shim struct could
be used to deal not only with shim pe files, but also with grub pe
files, for example.

This commit divices the shim module into the shim and pefile modules and
renames the Shim struct to PeFile. That way, it won't be as smelly if we
use the former Shim struct to load the grub binary and deal with it.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>