efivars: Refactoring efivar specific logic back
There was some logic from efivar that was initially written in the
uefi::efivar submodule, expecting to have other loader implementations.
However, for some reason, that logic was moved into the uefi module,
which didn't make sense at all. Now that we are getting ready for other
uefi variable input types, we need to take the efivar specific code into
where it belongs.

This commit also implements a trait to describe how secure boot variable
loaders should behave.

It also derives the clone and debug traits for the EFIVarLoader to avoid
some of the borrow/move issues that appeared on main.rs after the
refactoring.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>