efivars: Support efivars when secureboot is disabled
When secureboot is disabled the EFIVAR files related to PK, KEK db and
dbx aren't created if those variables are empty. That's expected, not a
problem. In that case the tpm events are logged as well, showing that
the variables are empty (data len = 0).

We can't just return an empty vec from load_uefi_var_data whenever we
can't open the file. We have to make sure that the problem is that the
file is not found, even if the directory in which is stored exists. We
should still panic for any other problem opening the file, such as wrong
permissions, directory not existing, etc.

This commit also adds a little fix to certs::get_db_certs_raw to make
sure that it doesn't run on empty vectors. If an empty vector is passed,
which is ok, it will return an empty certificate vector instead of an
error or even panicing.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>