pefile: Support calculating authenticode for vmlinuz
It seems the lief package is not able to handle vmlinuz images properly.
Apart from raising a bunch of errors and warnings when tried to load
such an image, it will fail to compute the authenticode.

This commit adds a workaround for vmlinuz images, for which it will
compute the digest of their signatures' content_info. For the rest of
images it will remain the same, but with a direct way of accessing the
computation through &self.authenticode, instead of
&self.image.authentihash.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>