pcr4: Parametrize vmlinuz path
Let the user define where the program needs to find the vmlinuz image
path. It defaults to the path in which they can be found in an fcos
container image.

This commit also adds a really naive piece of logic that will apply glob
on the input to find the vmlinuz image. That way we won't depend on the
version to write the command line.

Note that it is so naive that it will return just the first finding.
This could, or could not be the latest kernel version available under
that dir-tree. Beware that the selection algorithm is rather arbitrary.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>