justfile: Run test-container on the compute-pcrs
We are now able to build a compute-pcrs container image based on fedora.
Instead of running the tests on the target container image we want to
test (say fcos,scos,rhcos... whatever), we would rather run the binary
on the container image we just built, and mount the target container
image as a volume.

That way, we will be able to avoid many of the compatibility issues we
would face when trying to run a binary built on fedora, on any of the
other target container images.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>