justfile: Avoid compatibility issues in test-container
If an image other than fcos 42 was used, just test-container could fail
due to possible compatibility issues between the system in which
compute-pcrs was compiles and the container.

To avoid that, we propose running the test on a fedora:latest container
image. This was chosen just because, but it could be updated to a proper
container image containing the compiled compute-pcrs binary as soon as
we are able to build it.

Then, we mount the target image (fcos, rhcos...) that we want to test
to the container under /var/srv/image. Then, compute-pcrs can look for
vmlinuz, esp and other dependencies there instead.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>