Containerfile: Build the image from fedora images
Instead of using debian-based images, use fedora-based ones which will
be probably closer to where we will finally be using the tool.

Use the github's buildroot image to build the binary, which already
contains all the packages needed for that. Then copy the binary into a
fresh fedora container image.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>