justfile: Support fetching target container images from urls
Some of the container images we will have to deal with won't be directly
accessible on a registry. We will probably need to fetch them from a url
storing an ostree ociarchive.

That's why this commit introduces some changes in the justfile, so we
can download, and load container images from endpoints hosting
ociarchive files.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>