justfile: Run tests on the built container image
We are now able to build a container image with the compute-pcrs binary.
Switch the test logic so we follow the same approach followed in the
test-container test: the target container image is mounted onto the
compute-pcrs container, and the paths are passed pointing into the right
directions in the target container.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>