justfile: Get rid of get-test-data
Now tests will run on the compute-pcrs container image. The target image
will be mounted onto that one. We don't need to fetch the vmlinuz,
shim... for some tests anymore. No test depend on get-test-data right
now.

This commit just removes that piece of code from the justfile in favor
of the new proposed way of testing the binary against different OS
images.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>