justfile: Configure tests based on image osinfo
Aiming into running tests against a bunch of different Core OS images
and distros, we need to configure the tests based on the OS information.
Test configuration in this scenario means where tests should look for
the efi, mok variables and test-fixtures.

The available information is the one we can find under /etc/os-release.
This commit adds a new recipe onto which tests will depend into by
default, which extracts the os-release contents and puts them into a
file in /tmp on the host. This is done as "just" does not let passing
environment variables from one recipe into another. That is, we need an
intermediate file to hold the contents for a little bit, which we will
read afterwards, right before running the actual tests.

Tests are configured over ID, VERSION_ID and OSTREE_RELEASE. The files
under test-fixtures followed a tree directory naming rule that was hard
to follow based on the available info. This commit also moves the files
under test-fixtures so they can be easily found by the tests.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>