.github: Run integration tests against rhcos too
We want to run integration tests against a bunch of different images.
Previous commits prepare the justfile to be able to do that through the
"image" variable of the justfile.

This commit adds a matrix in the .github ci integration test job so it
goes through fcos, as well as an rhcos image.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>