cli: Only accept path to target rootfs
Paths to the kernel directory and the ESP (containing both grub and shim
binaries) were received as separate arguments. However, this tools is
run against a single container image at a time, meaning that we could
easily compute the paths to the kernel dir and the ESP dir just by
knowing were the rootfs of that container image is.

This commit introduces the simple logic that makes that possible, and
introduces the new --rootfs argument to the CLI.

Note that this commit introduces a breaking change, as it also removes
the old --esp and --kernels arguments from the CLI and the lib's API in
favor of the new rootfs argument.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>