cli: Only accept path to target rootfs
Paths to the kernel directory and the ESP (containing both grub and shim
binaries) were received as separate arguments. However, this tool is
run against a single container image at a time, meaning that we could
easily compute the paths to the kernel dir and the ESP dir just by
knowing were the rootfs of that container image is.

This commit introduces the simple logic that makes that possible, and
introduces the new --rootfs argument to the CLI.

Note that this commit only changes the CLI, not the library interface.
Now, users of the library can use the RootFSTree abstraction added in a
previous commit to compute the vmlinuz and esp paths, and then call the
predict functions. This is done as it provides greater flexibility for
library consumers (in favour of comfort) and changing it to only
accepting the rootfs path would have meant a breaking change instead.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>