justfile: Divide test results per host/cloud platform
efivars in reference values are host OS or cloud platform dependent, not
guest OS dependent. See

    https://github.com/confidential-clusters/reference-values/pull/4

This commit updates the test-fixture hierarchy, adds a little README
explaining it and updates the justfile to find the right paths.

Consider that we are just testing fedora-42 as a host platform for now,
but we would probably soon add tests/test-fixtures for other platforms
as well (other cloud platforms, centos+qemu-kvm/ovmf...).

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>