lib: Keep hashes as Vec<u8> in memory
Part and PCR hashes were modelled as Strings. This made it a bit more
difficult than it should to handle them in memory (or reuse them later).

This commit uses the serde_as library to represent hash values as
Vec<u8> in memory in favour of comfort but still serialize them into
hex-encoded Strings for ease of comprehension.

Note this is a breaking API change that could affect the way the
cocl-operator consumes the library.

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>