TPMEvents: Remove MixModel
Knowing the group models outside of the execution is not needed. It
doesn't add any value to the TPMEvents. In other words, the tpm event
group-id relationships are static. They do not change over time. If they
do, it will be a bug, which will be easier to fix by just changing a
line of code than updating many json annotations and files wherever
they are used.

This commit removes the TPMEventMixModel struct from there, adds a
function to set the tpm-id->group relationships, and makes the TPMEvent
struct to only hold the Event ID. It also renames the field from mix, to
id.

Also runs some updates to the code where "mix" or anything related was
referenced instead of the brand new "id".

Signed-off-by: Beñat Gartzia Arruabarrena <bgartzia@redhat.com>